Change float type variables to double.
We don't need float as we are not too worried about
space savings and we want to avoid rounding errors.

+++ src/main/java/seng201/team56/models/RoundDifficulty.java
@@ -1,4 +1,4 @@
 package seng201.team56.models;
 
-public record RoundDifficulty(double trackDistance, int numCarts, int cartMinSize, int cartMaxSize, float cartMinSpeed,
-                              float cartMaxSpeed) {}
+public record RoundDifficulty(double trackDistance, int numCarts, int cartMinSize, int cartMaxSize, double cartMinSpeed,
+                              double cartMaxSpeed) {}
